Bugfix only assign direction for attack and move
In replay games, guarding doesn't have a target, I took this to mean
that only attack and move actions have a target and used that to
differentiate the types.

In a live game, the guard action has a target of your own square, which
resulted in an undefined direction of {0, 0} and a crash.